facial-recognition-app/backend/requirements.txt
Flask
tensorflow
numpy
Pillow
opencv-python
scikit-learn
flask-cors